* commit 'dde9b7e5db377d3990f8e5bfefec801ca41a55f6':
  SW-22817 - Fix category doctype
  SW-22817 - Fix all the typehints

+++ engine/Shopware/Components/HttpCache/DynamicCacheTimeService.php
@@ -72,7 +72,7 @@ class DynamicCacheTimeService implements CacheTimeServiceInterface
         /** @var InvalidationDateInterface $dateProvider */
         foreach ($this->invalidationDateProvider as $dateProvider) {
             if ($dateProvider->supportsRoute($action) && $invalidationDate = $dateProvider->getInvalidationDate($request)) {
-                $difference = $invalidationDate->format('U') - (new DateTime())->format('U');
+                $difference = (int) $invalidationDate->format('U') - (int) (new DateTime())->format('U');
 
                 return $difference > 0 && $difference < $defaultInvalidationTime ? $difference : $defaultInvalidationTime;
             }
